feat: making the nfInstanceId configurable
Initially the NF created at each boot a new uuid. This makes it difficult to retrieve the information afterwards. This commit propose is to add the nfInstanceId in the configuration as optional so that if needed, it can be tracked more easily.
If no nfInstanceId specified, it will still be generated at each boot and give more control to users that might need it (for kube deployment, other checks, better identification)

Also using the uuidv4 validator to match spec 3GPP TS 29.510 version 17.6.0 Release 17 5.2.2.2.2 mentioning that UUID should be in the version 4 format